requirements.txt

asgiref==3.8.1
django==5.1
djangorestframework==3.15.2
pytz==2024.1
sqlparse==0.5.1
